backend/requirements.txt
# Django and REST Framework
Django==5.0.2
djangorestframework==3.14.0
drf-nested-routers==0.93.4
django-cors-headers==4.3.1
django-filter==23.5

# Database
psycopg2-binary==2.9.9
django-environ==0.11.2

# Authentication
djangorestframework-simplejwt==5.3.0
django-allauth==0.60.1

# File handling and parsing
python-docx==1.1.0
PyPDF2==3.0.1
pdfplumber==0.10.3
Pillow==10.2.0

# Async tasks
celery==5.3.4
redis==5.0.1

# Storage
django-storages==1.14.2
boto3==1.34.34

# PDF generation
reportlab==4.0.7
weasyprint==60.2

# Text processing and NLP
spacy==3.7.2
nltk==3.8.1
python-Levenshtein==0.23.0

# Security
argon2-cffi==23.1.0
cryptography==41.0.8

# Development and testing
pytest==7.4.4
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.4.1

# Monitoring
sentry-sdk==1.40.4

# Utilities
python-dateutil==2.8.2
requests==2.31.0
